# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "B7Store API",
        "description": "API completa para o sistema B7Store - E-commerce",
        "contact": {
            "email": "suporte@b7store.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor Local"
        },
        {
            "url": "https://api.b7store.com",
            "description": "Servidor de Produção"
        }
    ],
    "paths": {
...
        "/api/product": {
            "get": {
                "tags": [
                    "Product"
                ],
                "summary": "Listar produtos",
                "description": "Retorna lista de produtos com filtros opcionais",
                "operationId": "5cc57bec35d0f2045983cc8dcd00eccd",
                "parameters": [
                    {
                        "name": "orderby",
                        "in": "query",
                        "description": "Ordenar por: views, selling, price",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Limite de resultados",
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    },
                    {
                        "name": "metadata",
                        "in": "query",
                        "description": "Filtro JSON de metadados",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de produtos"
                    },
                    "500": {
                        "description": "Erro ao listar produtos"
                    }
                }
            }
        },
        "/api/product/{id}": {
            "get": {
                "tags": [
                    "Product"
                ],
                "summary": "Detalhes do produto",
                "description": "Retorna informações completas de um produto e incrementa contador de visualizações",
                "operationId": "2768d88590e39a7ded1d0032e64dded7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalhes do produto"
                    },
                    "500": {
                        "description": "Erro ao listar produto"
                    }
                }
            }
        },
        "/api/product/{id}/related": {
            "get": {
                "tags": [
                    "Product"
                ],
                "summary": "Produtos relacionados",
                "description": "Retorna produtos da mesma categoria, excluindo o produto atual",
                "operationId": "b1d0ce6f696eceaab1200fa9dd0ae7b7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "example": 4
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de produtos relacionados"
                    },
                    "500": {
                        "description": "Erro ao listar produtos relacionados"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Token de autenticação via Laravel Sanctum. Use: Bearer {seu-token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}